runtime: tidying up and docs
