+++ packages/runtime/README.md
@@ -1,20 +1,24 @@
 ## Runtime
 
-A runtime for running openfn jobs and reporting on their status.
+A runtime for running openfn workflows and reporting on their status. The runtime will return a serializable state object when the workflow is completed.
 
-The runtime will load an array of operations from a module and execute them in series.
+A single expression can be passed to the runtime, which will be wrapped into a single-node wowrkflow.
 
-An operation is a function which takes state as input and returns state, or a promise resolving to state, as output.
+The runtime is designed to be wrapped into a Runtime Manager, which is responsible for compiling expressions into executable code, triggering runs (ie calling the runtime), and returning state.
+
+Each job/expression in a workflow must export an array of operations, which will be executed in series.
+
+An operation is a function which takes state as input and returns state (or a promise resolving to state) as output.
 
 ```js
 run([(state) => state]);
 ```
 
-The compiler can be used to convert job DSL into an compatible ESM module (the runtime does not do this automatically).
+The compiler can be used to convert the javascript-like "job DSL" into an compatible ESM module (the runtime does NOT do this automatically).
 
 ## Basic Usage
 
-The runtime should be passed the source for a single job (as a string, as a module which exports an array of functions.)
+The runtime should be passed a workflow JSOn object, or a single job (as a string or as a module which exports an array of functions).
 
 ```js
 import { readFile } from 'node:fs/promises';
@@ -22,7 +26,8 @@ import run from '@openfn/runtime';
 
 const job = await readFile('expression.js', 'utf8');
 const initialState = {};
-const { data } = await run(source, initialState);
+const options = { logger };
+const { data } = await run(source, initialState, options);
 ```
 
 See the `test` folder for more usage examples.
@@ -32,6 +37,48 @@ The runtime provides no CLI. Use packages/cli (devtools) for this.
 ## Experimental VM Args
 
 For the runtime to work, the parent process needs `--experimental-vm-modules` be passed. You may also want to pass `--no-warnings` to suppress annoying console warnings.
+
+## Execution Plans
+
+The runtime can accept an Execution Plan (or workflow) as an input. 
+
+This defines a graph of of jobs (expressions) to run in sequence. Each node in the graph is a job, and contains a set of edges which tell the runtime what to execute next.
+
+The runtime will return the final state when there is nothing left to execute.
+
+A workflow looks like this:
+```
+{
+  start: 'a',
+  jobs: [{
+    id: 'a',
+    expression: "source or path",
+    data: { /* default data */ },
+    configuration: { /* credentials */ },
+    next: {
+      'b': true, // edge to another job
+      'c': { condition: "state.data.age > 18", // conditional edge to another job
+    }    
+    adaptor: "common", // it's complicated
+  }]
+}
+```
+State and start node can be passed to the runtime as inputs.
+
+If no start node is provided, the first job in the jobs array will run first.
+
+Ids are technically optional, but needed if an edge (or start) wants to refer to a node.
+
+The runtime itself does not use the `adaptor` key, as it expects jobs to be compiled with imports. As with expressions, it's the runtime manager's job to compile expressions and ensure dependencies are available.
+
+See src/types.ts for a full definition of an execution plan, and `test/runtime.test.ts` for examples.
+
+At the time of writing, exectuion plans have some restrictions:
+* Jobs execute in series (but parallisation can be simulated)
+* A job can only have one input node (`a -> z <- b` is not allowed) 
+* Jobs cannot have circular references (`a -> b -> a` is not allowed)
+
+Support for more complex plans will be introduced later.
 
 ## Building
 
@@ -58,25 +105,28 @@ $ pnpm test:watch
 
 ## Runtime Design
 
-The runtime's job is to take a pipline of operations and execute them in series.
+The runtime's job is to take one or more jobs and execute them in series.
 
-The runtime should:
+Each job, in turn, is a pipeline of operations.
 
-- Accept a pipleline as an array of functions or a stringified ESM module
-- Validate the input string to ensure there's no security concerns (like blacklisted imports)
-- Execute the pipeline in a safe environment (with some utilities and overrides provided)
-- Ensure that the state object is not mutated between jobs
-- Return a promise and event-emitted (with a `on(event)` function)
-- Emit lifecycle events for the job pipeline
-- Resolve to a state object
-- Load runtime dependencies from explicit paths or a local repo
+The runtime's repsonsibilities include:
+
+- Accepting a workflow as JSON or a job as as string
+- Executing all jobs in a safe environment (with some utilities and overrides provided)
+- Ensuring that the state object is not mutated between jobs
+- Emitting lifecycle events for the job pipeline
+- Maintaining a repo of node modules which are available to jobs
+- Loading runtime dependencies from explicit paths (passed in) or a local repo
+- Resolving to a serializable state object
 
 The runtime should not:
 
-- Compile its input jobs (although it will validate using the compiler)
+- Compile its input jobs (although it may validate using the compiler)
 - Do any disk I/O
-- Do any thread/process management (see the runtime manager)
+- Do any thread/process management
 - Auto install any dependencies
+
+These are all the responsibilities of a runtime manager (like the CLI).
 
 ## Module Loading
